Add in-memory initial package.
