+++ src/Tests/SenseNet.ODataTest.WebApp/Program.cs
@@ -31,7 +31,6 @@ namespace SenseNet.ODataTest.WebApp
             }))
             {
                 SnTrace.EnableAll();
-                InitializeDatabase();
                 host.Run();
             }
         }
@@ -39,31 +38,5 @@ namespace SenseNet.ODataTest.WebApp
         public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
             WebHost.CreateDefaultBuilder(args)
                 .UseStartup<Startup>();
-
-        private static void InitializeDatabase()
-        {
-            var path = Path.GetFullPath(Path.Combine(AppDomain.CurrentDomain.BaseDirectory,
-                @"..\..\..\..\..\nuget\snadmin\install-services-core\manifest.xml"));
-            string manifestXml;
-            using (var reader = new StreamReader(path))
-                manifestXml = reader.ReadToEnd();
-            var xml = new XmlDocument();
-            xml.LoadXml(manifestXml);
-            var root = xml.DocumentElement;
-
-            var package = new Package
-            {
-                ComponentId = root.SelectSingleNode("Id").InnerText,
-                Description = root.SelectSingleNode("Description").InnerText,
-                PackageType = Enum.Parse<PackageType>(root.Attributes["type"].Value),
-                ReleaseDate = DateTime.Parse(root.SelectSingleNode("ReleaseDate").InnerText),
-                ComponentVersion = Version.Parse(root.SelectSingleNode("Version").InnerText),
-                ExecutionDate = DateTime.UtcNow,
-                ExecutionResult = ExecutionResult.Successful,
-                Manifest = manifestXml
-            };
-            var provider = DataStore.GetDataProviderExtension<IPackagingDataProviderExtension>();
-            provider.SavePackageAsync(package, CancellationToken.None).ConfigureAwait(false).GetAwaiter().GetResult();
-        }
     }
 }
